fix: rename sitemap app to web_client
