+++ terraso_backend/config/urls.py
@@ -19,7 +19,7 @@ from django.contrib import admin
 from django.contrib.sitemaps.views import sitemap
 from django.urls import include, path
 
-from apps.sitemap.sitemap import WebClientSitemap
+from apps.web_client.sitemap import WebClientSitemap
 
 # OAuth2 provider endpoints
 oauth2_endpoint_views = [
